requirements.txt
opencv-python>=4.9.0
pytesseract>=0.3.10
numpy>=1.26.0
pillow>=10.0.0
paho-mqtt>=2.1.0
boto3>=1.34.0
python-dotenv>=1.0.1
PyYAML>=6.0.1
